% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gtkNotebookInsertPageWithCloseButton}
\alias{gtkNotebookInsertPageWithCloseButton}
\title{Insert page in notebook}
\usage{
gtkNotebookInsertPageWithCloseButton(object, child, label.text = "",
  position = -1)
}
\description{
Insert page in notebook
}
\keyword{internal}
