% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_gui.R
\name{IAEN}
\alias{IAEN}
\title{Introduction to the analysis of ecological networks}
\usage{
IAEN(h, ...)
}
\arguments{
\item{h}{Not usable}

\item{...}{Not usable}
}
\value{
IAEN is a graphic interface
}
\description{
The IAEN package is a tool that provides a graphical user interface (GUI), allows data management performing different types of analyses used in ecological networks. Includes visualization and editing of graphics, methods for binary, weighted and bipartite ecological networks. You can also simulate binary networks using some classic models, using the networks topology, n number of repetitions can be produced, and use functions employed in networks of small worlds.

You can learn more about this package at:
https://github.com/ihuesca
}
\details{
IAEN is a package with a graphical interface dedicated to analysis of ecological networks.

Note: IAEN is free software and comes with ABSOLUTELY NO WARRANTY.
}
\examples{
\dontrun{
##Install package
library(IAEN)
##Call the package
IAEN()
}

}
\references{
Cohen, J. E., & Newman, C. M. (1985). A stochastic theory of community food webs I.
Models and aggregated data. Proceedings of the Royal society of London. Series B.
Biological sciences, 224(1237), 421-448.

Dame, R. F., & Patten, B. C. (1981). Analysis of energy flows in an intertidal oyster reef.
Marine Ecology Progress Series, 5(2), 115-124.

Dormann, C. F. (2011). How to be a specialist? Quantifying specialisation in pollination
networks. Network Biology, 1(1), 1-20.

Erds, P., & Rényi, A. (1960). On the evolution of random graphs. Publ. Math. Inst.
Hungar. Acad. Sci, 5, 17-61.

Goldwasser, L., & Roughgarden, J. (1993). Construction and Analysis of a Large Caribbean
Food Web: Ecological Archives E074-001. Ecology, 74(4), 1216-1233.

GTK+ (2007). The Gimp Tool Kit. url: http://www.gtk.org/.

Lawrence, M., & Temple Lang, D. (2010). RGtk2: A graphical user interface toolkit for R.
Journal of Statistical Software, 37(8), 1-52.

Ulanowicz, R. E., & Puccia, C. J. (1990). Mixed trophic impacts in ecosystems.
Coenoses, 7-16.

Watts, D. J., & Strogatz, S. H. (1998). Collective dynamics of ‘small-world’networks.
Nature, 393(6684), 440.

Williams, R. J., & Martinez, N. D. (2000). Simple rules yield complex food webs.
Nature, 404(6774), 180.
}
\author{
Luis Gerardo Abarca-Areanas, Israel Huesca-Dominguez
}
